pyproject.toml
[project]
name = "temp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "customtkinter>=5.2.2",
    "imagehash>=4.3.2",
    "pillow>=12.1.1",
    "pyqt6>=6.10.2",
]
